+++ src/components/atoms/Typography/typographyVariants.ts
@@ -1,0 +1,209 @@
+export const variant: any = {
+  body_text_1: {
+    size: "1.1rem",
+    lineHeight: "1.8rem",
+    spacing: "0.3px",
+    tablet: {
+      size: "1.2rem",
+      lineHeight: "2rem",
+      spacing: "0.4px",
+    },
+  },
+  body_text_2: {
+    size: "1.2rem",
+    lineHeight: "2rem",
+    spacing: "0.4px",
+    tablet: {
+      size: "1.4rem",
+      lineHeight: "2.4rem",
+      spacing: "0",
+    },
+  },
+  body_text_3: {
+    size: "1.3rem",
+    lineHeight: "2rem",
+    spacing: "0.4px",
+    tablet: {
+      size: "1.5rem",
+      lineHeight: "2.2rem",
+      spacing: "2.15px",
+    },
+    laptop: {
+      size: "1.7rem",
+      lineHeight: "3rem",
+      spacing: "2.15px",
+    },
+  },
+  body_text_4: {
+    size: "1.4rem",
+    lineHeight: "2.4rem",
+    spacing: "0.4px",
+  },
+  body_text_5: {
+    size: "1.4rem",
+    lineHeight: "2.2rem",
+    spacing: "0.4px",
+    tablet: {
+      size: "1.2rem",
+      lineHeight: "2rem",
+      spacing: "0.4px",
+    },
+  },
+  body_text_6: {
+    size: "1.5rem",
+    lineHeight: "2.2rem",
+    spacing: "0.35",
+    tablet: {
+      size: "1.8rem",
+      lineHeight: "2.4rem",
+      spacing: "0",
+    },
+  },
+  body_text_7: {
+    size: "1.6rem",
+    lineHeight: "2.4rem",
+    spacing: "0",
+  },
+  body_text_8: {
+    size: "1.6rem",
+    lineHeight: "2.4rem",
+    spacing: "0",
+    tablet: {
+      size: "1.7rem",
+      lineHeight: "3rem",
+      spacing: "0.4px",
+    },
+  },
+  body_text_9: {
+    size: "2rem",
+    lineHeight: "2.8rem",
+    spacing: "0",
+    tablet: {
+      size: "2.4rem",
+      lineHeight: "3.8rem",
+      spacing: "0.3",
+    },
+  },
+  text_uppercase_1: {
+    transform: "uppercase",
+    size: "1.2rem",
+    lineHeight: "2rem",
+    spacing: "0.4px",
+  },
+  text_uppercase_2: {
+    transform: "uppercase",
+    size: "1.4rem",
+    lineHeight: "2.4rem",
+    spacing: "1.75px",
+    tablet: {
+      size: "1.7rem",
+      lineHeight: "3rem",
+      spacing: "2.15px",
+    },
+  },
+  heading_8: {
+    size: "1.3rem",
+    lineHeight: "2rem",
+    spacing: "0.4px",
+    tablet: {
+      size: "2rem",
+      lineHeight: "2.8rem",
+      spacing: "0",
+    },
+  },
+  heading_7: {
+    size: "1.4rem",
+    lineHeight: "2.4rem",
+    spacing: "0.4px",
+    tablet: {
+      size: "2rem",
+      lineHeight: "2.8rem",
+      spacing: "0.4px",
+    },
+    laptop: {
+      size: "2.6rem",
+      lineHeight: "4rem",
+      spacing: "0.35px",
+    },
+  },
+  heading_6: {
+    size: "1.7rem",
+    lineHeight: "3rem",
+    spacing: "0.4px",
+    tablet: {
+      size: "2rem",
+      lineHeight: "2.8rem",
+      spacing: "0",
+    },
+  },
+  heading_5: {
+    size: "2.4rem",
+    lineHeight: "3.8rem",
+    spacing: "0.3px",
+    tablet: {
+      size: "3.5rem",
+      lineHeight: "4.2rem",
+      spacing: "0.45px",
+    },
+    laptop: {
+      size: "3.7rem",
+      lineHeight: "6.2rem",
+      spacing: "0",
+    },
+  },
+  heading_4: {
+    size: "2.4rem",
+    lineHeight: "3.8rem",
+    spacing: "0.3px",
+    tablet: {
+      size: "3.5rem",
+      lineHeight: "4.2rem",
+      spacing: "0.45px",
+    },
+    laptop: {
+      size: "4.4rem",
+      lineHeight: "6.2rem",
+      spacing: "0",
+    },
+  },
+  heading_3: {
+    size: "2.4rem",
+    lineHeight: "3.8rem",
+    spacing: "0.3px",
+    tablet: {
+      size: "3.5rem",
+      lineHeight: "4.2rem",
+      spacing: "0.45px",
+    },
+    laptop: {
+      size: "4.6rem",
+      lineHeight: "6.2rem",
+      spacing: "0",
+    },
+  },
+  heading_2: {
+    size: "2.6rem",
+    lineHeight: "4rem",
+    spacing: "0.35px",
+    tablet: {
+      size: "3.7rem",
+      lineHeight: "4.8rem",
+      spacing: "0.35px",
+    },
+    laptop: {
+      size: "4.6rem",
+      lineHeight: "6.2rem",
+      spacing: "0",
+    },
+  },
+  heading_1: {
+    size: "3.5rem",
+    lineHeight: "4.2rem",
+    spacing: "0.45px",
+    laptop: {
+      size: "5.7rem",
+      lineHeight: "6.8rem",
+      spacing: "0",
+    },
+  },
+};
